+++ services/v1/tag.go
@@ -1,6 +1,8 @@
 package v1
 
 import (
+	"errors"
+
 	"github.com/atticuss/chefconnect/models"
 	"github.com/atticuss/chefconnect/services"
 )
@@ -27,6 +29,17 @@ func (s *v1Service) GetTag(callingUser *models.User, id string) (*models.Tag, *s
 
 // CreateTag handles the business logic when a client creates a new tag
 func (s *v1Service) CreateTag(callingUser *models.User, tag *models.Tag) (*models.Tag, *services.ServiceError) {
+	authorized := false
+	for _, role := range callingUser.Roles {
+		if role.Name == services.Admin {
+			authorized = true
+		}
+	}
+
+	if !authorized {
+		return tag, &services.ServiceError{Error: errors.New("unathorized"), ErrorCode: services.NotAuthorized}
+	}
+
 	tag, err := s.TagRepository.Create(tag)
 
 	if err != nil {
@@ -38,6 +51,17 @@ func (s *v1Service) CreateTag(callingUser *models.User, tag *models.Tag) (*model
 
 // UpdateTag handles the business logic when a client updates a tag
 func (s *v1Service) UpdateTag(callingUser *models.User, tag *models.Tag) (*models.Tag, *services.ServiceError) {
+	authorized := false
+	for _, role := range callingUser.Roles {
+		if role.Name == services.Admin {
+			authorized = true
+		}
+	}
+
+	if !authorized {
+		return tag, &services.ServiceError{Error: errors.New("unathorized"), ErrorCode: services.NotAuthorized}
+	}
+
 	tag, err := s.TagRepository.Update(tag)
 
 	if err != nil {
@@ -49,6 +73,17 @@ func (s *v1Service) UpdateTag(callingUser *models.User, tag *models.Tag) (*model
 
 // DeleteTag handles the business logic when a client deletes a tag
 func (s *v1Service) DeleteTag(callingUser *models.User, id string) *services.ServiceError {
+	authorized := false
+	for _, role := range callingUser.Roles {
+		if role.Name == services.Admin {
+			authorized = true
+		}
+	}
+
+	if !authorized {
+		return &services.ServiceError{Error: errors.New("unathorized"), ErrorCode: services.NotAuthorized}
+	}
+
 	err := s.TagRepository.Delete(id)
 	if err != nil {
 		return &services.ServiceError{Error: err}
